requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.7
cfgv==3.3.1
distlib==0.3.6
et-xmlfile==1.1.0
Faker==17.0.0
filelock==3.9.0
flake8==6.0.0
flake8-builtins==2.1.0
flake8-comments==0.1.2
identify==2.5.17
kombu==4.6.11
mccabe==0.7.0
mypy==1.1.1
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
pep8-naming==0.13.3
platformdirs==3.0.0
pre-commit==3.0.4
pycodestyle==2.10.0
pycryptodome==3.17
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
vine==1.3.0
virtualenv==20.19.0
